requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.8
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
olefile==0.46
Pillow==7.1.2
requests==2.23.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.9
webptools==0.0.2
Werkzeug==1.0.1
